pyproject.toml
[project]
name = "amazon-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "faker>=40.12.0",
    "kafka-python>=2.3.0",
    "openpyxl>=3.1.5",
    "pandas>=3.0.2",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
]
